requirements.txt
streamlit==1.33.0
pandas==2.2.1
numpy==1.26.0
nltk==3.8.1
scikit-learn==1.4.1
textblob==0.17.1
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.5.0
firebase-admin==6.5.0
matplotlib==3.8.3
regex==2023.12.25 